add obstacle (a moving cylinder) between the hole and the ball, still DOT warning
